requirements.txt
Flask>=2.2
pandas>=2.0
numpy>=1.24
scikit-learn>=1.2
joblib>=1.2
tensorflow>=2.11.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0
